requirements.txt
streamlit==1.39.0
mysql-connector-python==8.4.0
pandas==2.2.2
python-dotenv==1.0.1